(FIXME: REVERT THIS!) Remove gradle-api from the runtime classpath
This is a workaround against class ambiguity: the dependency on the
plugins from an included build brings the Gradle API JAR on the
classpath. That has SLF4J bundled, which clashes with the one used
in this project.

Revert this commit once we switch to a dependency on published plugins.